完善了权限管理和部门管理

+++ IntakeSystem/src/main/java/com/student_online/IntakeSystem/service/StationService.java
@@ -16,46 +16,81 @@ public class StationService {
     @Autowired
     private StationMapper stationMapper;
 
-    public ResponseEntity<Result> createStation(@NotNull Station station) {
-        String name=station.getName();
-        if(stationMapper.getStationByName(name)!=null)return ResponseUtil.build(Result.error(409,"该模块已存在"));
-        else{
-            stationMapper.createStation(station);
-            return ResponseUtil.build(Result.ok());
+    public ResponseEntity<Result> createStation(@NotNull Station station,int uid) {
+        try {
+            String name = station.getName();
+            int pid = station.getPId();
+            if (stationMapper.getStationByName(name) != null)
+                return ResponseUtil.build(Result.error(409, "该模块已存在"));
+            else {
+                PermissionService permissionService = new PermissionService();
+                if (permissionService.isPermitted(pid, uid)) {
+                    stationMapper.createStation(station);
+                    return ResponseUtil.build(Result.ok());
+                } else return ResponseUtil.build(Result.error(401, "无权限"));
+            }
+        }catch (Exception e) {
+            return ResponseUtil.build(Result.error(400, "创建失败"));
         }
 
     }
 
-    public ResponseEntity<Result> updateStation(@NotNull Station station) {
-        stationMapper.updateStation(station);
-        return ResponseUtil.build(Result.ok());
+    public ResponseEntity<Result> updateStation(Station station,int uid) {
+        try {
+            PermissionService permissionService = new PermissionService();
+            int stationId = station.getId();
+            if (permissionService.isPermitted(stationId, uid)) {
+                stationMapper.updateStation(station);
+                return ResponseUtil.build(Result.ok());
+            } else return ResponseUtil.build(Result.error(401, "无权限"));
+        }catch (Exception e) {
+            return ResponseUtil.build(Result.error(400, "更新失败"));
+        }
+
     }
 
-    public ResponseEntity<Result> deleteStation( int stationId) {
+    public ResponseEntity<Result> deleteStation( int stationId,int uid) {
         try {
-            stationMapper.deleteStationById(stationId);
-            return ResponseUtil.build(Result.ok());
+            PermissionService permissionService = new PermissionService();
+            if(permissionService.isPermitted(stationId, uid )) {
+                stationMapper.deleteStationById(stationId);
+                return ResponseUtil.build(Result.ok());
+            }
+            else return ResponseUtil.build(Result.error(401,"无权限"));
         }catch (Exception e){
             return ResponseUtil.build(Result.error(404, "模块不存在"));
         }
     }
 
     public ResponseEntity<Result> getStationById( int stationId) {
-        Station station=stationMapper.getStationById(stationId);
-        if(station==null)return ResponseUtil.build(Result.error(404,"未找到该模块"));
-        else return ResponseUtil.build(Result.success(station,"返回模块"));
+        try {
+            Station station = stationMapper.getStationById(stationId);
+            if (station == null) return ResponseUtil.build(Result.error(404, "未找到该模块"));
+            else return ResponseUtil.build(Result.success(station, "返回模块"));
+        }catch (Exception e) {
+            return ResponseUtil.build(Result.error(400, "获取失败"));
+        }
     }
 
     public ResponseEntity<Result> getStationByName(@NotNull String name) {
-        Station station=stationMapper.getStationByName(name);
-        if(station==null)return ResponseUtil.build(Result.error(404,"未找到该模块"));
-        else return ResponseUtil.build(Result.success(station,"返回模块"));
+        try {
+            Station station = stationMapper.getStationByName(name);
+            if (station == null) return ResponseUtil.build(Result.error(404, "未找到该模块"));
+            else return ResponseUtil.build(Result.success(station, "返回模块"));
+        }catch (Exception e) {
+            return ResponseUtil.build(Result.error(400, "获取失败"));
+        }
+
     }
 
     public ResponseEntity<Result> getStationByParentId( int pid) {
-        List<Station> result=stationMapper.getStationByParentId(pid);
-        if(result.isEmpty())return ResponseUtil.build(Result.error(404,"未找到它的子模块"));
-        else return ResponseUtil.build(Result.success(result,"返回子模块"));
+        try {
+            List<Station> result = stationMapper.getStationByParentId(pid);
+            if (result.isEmpty()) return ResponseUtil.build(Result.error(404, "未找到它的子模块"));
+            else return ResponseUtil.build(Result.success(result, "返回子模块"));
+        }catch (Exception e) {
+            return ResponseUtil.build(Result.error(400, "获取失败"));
+        }
     }
 
 }
